% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashua.R
\docType{data}
\name{ashua}
\alias{ashua}
\title{River Ashuapmushuan Daily Flow and Level Data}
\format{
A data frame with 4389 rows and 3 variables:
\describe{
  \item{flow}{Flow rate of the river in cubic meters per second (m³/s)}
  \item{level}{Level of the river in meters (m)}
  \item{day}{Date of the reading (class \code{day})}
}
}
\source{
Environment and Climate Change Canada Historical Hydrometric Data, \url{https://wateroffice.ec.gc.ca/mainmenu/historical_data_index_e.html}
}
\usage{
ashua
}
\description{
This dataset contains 4389 daily readings of the flow rate and level of the river Ashuapmushuan in cubic meters per second (m³/s) and meters (m), respectively. Data covers the period from 22 March 1992 to 30 September 2007, with a few missing days.
}
\details{
The dataset was extracted from the Environment and Climate Change Canada Historical Hydrometric Data website on 26 July 2023. Data may contain some missing days within the period.
}
\examples{
# Load the ashua data
data(ashua)
# Basic summary of the flow and level variables
summary(ashua)

}
\keyword{datasets}
