% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_aliases.R
\name{lowercase_aliases}
\alias{lowercase_aliases}
\alias{ash}
\alias{lbfp}
\alias{glbfp}
\alias{ash_estimate}
\alias{lbfp_estimate}
\alias{glbfp_estimate}
\title{Lowercase aliases for the public API}
\usage{
ash(...)

lbfp(...)

glbfp(...)

ash_estimate(...)

lbfp_estimate(...)

glbfp_estimate(...)
}
\arguments{
\item{...}{Arguments passed to the corresponding uppercase function.}
}
\value{
The same object returned by the corresponding uppercase function.
}
\description{
These aliases follow common R naming style while preserving the original
uppercase function names used in earlier versions of the package.
}
