% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.glbfp_fit}
\alias{predict.glbfp_fit}
\title{Predict from GLBFP fit objects}
\usage{
\method{predict}{glbfp_fit}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A fitted object of class \code{"glbfp_fit"} or \code{"glbfp_grid"}.}

\item{newdata}{Optional matrix/data frame with points where prediction is
requested. For \code{"glbfp_fit"}, \code{newdata} is not supported.}

\item{...}{Additional arguments (unused).}
}
\value{
Numeric vector of predicted densities.
}
\description{
Prediction helper for fitted GLBFP objects.
}
