% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.glbfp_fit}
\alias{summary.glbfp_fit}
\title{Summarize GLBFP fit objects}
\usage{
\method{summary}{glbfp_fit}(object, ...)
}
\arguments{
\item{object}{A fitted object of class \code{"glbfp_fit"} or \code{"glbfp_grid"}.}

\item{...}{Additional arguments (unused).}
}
\value{
A list with class \code{"summary.glbfp_fit"} or \code{"summary.glbfp_grid"}.
}
\description{
Summarizes objects returned by \code{\link[=ASH]{ASH()}}, \code{\link[=LBFP]{LBFP()}}, \code{\link[=GLBFP]{GLBFP()}} and
their grid counterparts.
}
