% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASH.R
\name{ASH}
\alias{ASH}
\alias{print.ASH}
\title{Averaged Shifted Histogram (ASH) Estimator}
\usage{
ASH(
  x,
  data,
  b = compute_bi_optim(data, m = rep(1, ncol(data))),
  m = rep(1, ncol(data)),
  min_vals = apply(data, 2, min),
  max_vals = apply(data, 2, max)
)

\method{print}{ASH}(x, ...)
}
\arguments{
\item{x}{Numeric vector representing the point at which to estimate density.}

\item{data}{Matrix or data frame of observations (n rows, d columns).}

\item{b}{Numeric vector representing the bin width for each dimension. Default is optimal bi calculated using \code{compute_bi_optim}.}

\item{m}{Numeric vector representing the shift for each dimension. Default is 1 on each dimension.}

\item{min_vals}{Minimum value of initial grid. Default value is the minimum on each column.}

\item{max_vals}{Maximum value of initial grid. Default value is the maximum on each column.}

\item{object}{Object of class \code{"ASH"}.}

\item{...}{Additional arguments (unused).}
}
\value{
A list containing the estimated density at point `x` using the ASH estimator, the bin width vector `b`, the shift vector `m` and `x`
}
\description{
Computes the ASH density estimate at a single point `x`.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print(ASH)}: Print method for object of class \code{"ASH"}

}}
\examples{
# Example for ASH function with ashua data
x <- c(200, 30)
## with a selected bin width
b <- c(0.5, 0.5)
m <- c(1,1)
ASH(x, ashua[,-3], b, m)
## with optimal b and shift m=3
ASH(x, ashua[,-3], m = c(3,3))
}
\keyword{internal}
