% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLBFP-package.R
\docType{package}
\name{GLBFP-package}
\alias{GLBFP-package}
\title{GLBFP: General Linear Blend Frequency Polygon Density Estimation}
\description{
\code{GLBFP} provides one-point and grid-based density estimators based on
ASH, LBFP and GLBFP methodology for one and two-dimensional use cases,
with visualization helpers and bandwidth selection utilities.
}
\details{
Main entry points:
\itemize{
\item \code{\link[=ASH]{ASH()}}, \code{\link[=LBFP]{LBFP()}}, \code{\link[=GLBFP]{GLBFP()}}
\item \code{\link[=ASH_estimate]{ASH_estimate()}}, \code{\link[=LBFP_estimate]{LBFP_estimate()}}, \code{\link[=GLBFP_estimate]{GLBFP_estimate()}}
\item \code{\link[=compute_bi_optim]{compute_bi_optim()}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AurelienNicosiaULaval/GLBFP}
  \item Report bugs at \url{https://github.com/AurelienNicosiaULaval/GLBFP/issues}
}

}
\author{
\strong{Maintainer}: Aurélien Nicosia \email{aurelien.nicosia@mat.ulaval.ca}

}
