% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K(mi).R
\name{K_mi}
\alias{K_mi}
\title{Function that calculates \eqn{K(m_i)}}
\usage{
K_mi(mi)
}
\arguments{
\item{mi}{An integer representing the index value, typically a positive integer related to the number of bins or intervals in the context of shifted histograms.}
}
\value{
The computed value of \eqn{K(m_i)}, which is used in further statistical calculations or analyses.
}
\description{
This function computes the value of \eqn{K(m_i)}, which is an expression derived from the asymptotic properties of a statistical estimator related to the bandwidth selection in nonparametric density estimation.
}
\details{
The function implements the formula given by
\deqn{K(m_i) = \sqrt{\frac{1}{6} + \frac{1}{12m_i^2}} + \frac{4m_i^2 - 1}{6\sqrt{2}m_i} \log\left(\frac{\sqrt{3} + \sqrt{4m_i^2 + 2}}{\sqrt{4m_i^2 - 1}}\right).} This formula is derived from the theoretical analysis in the context of average shifted histograms. The terms in the formula correspond to corrections applied to the bandwidth selection to optimize statistical properties.
}
\examples{
\dontrun{
if(interactive()){
 # Calculate K for mi = 2
 K_mi(2)

 # Calculate K for mi = 5
 K_mi(5)
}
}
}
