% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBFP.R
\name{LBFP}
\alias{LBFP}
\alias{print.LBFP}
\title{LBFP Density Estimator at a Single Point}
\usage{
LBFP(
  x,
  data,
  b = compute_bi_optim(data, m = rep(1, ncol(data))),
  min_vals = apply(data, 2, min),
  max_vals = apply(data, 2, max)
)

\method{print}{LBFP}(x, ...)
}
\arguments{
\item{x}{Object from \code{LBFP} to print.}

\item{data}{Numeric matrix or data frame representing the dataset. Each row is an observation.}

\item{b}{Numeric vector representing the bin width for each dimension. Default is optimal bi calculated using \code{compute_bi_optim}.}

\item{min_vals}{Minimum value of initial grid (I_0(k)). Default value is the minimum on each column.}

\item{max_vals}{Maximum value of initial grid (I_0(k)). Default value is the maximum on each column.}

\item{...}{Additional arguments (unused).}
}
\value{
A list containing the estimated density at point x using the LBFP estimator, the bin width vector b, the standard deviation of the estimation and finally 95% confidence interval.
}
\description{
This function computes the LBFP (Linear Blend Frequency Polygon) estimate of the density
at a single given point. The LBFP is a continuous density estimator inspired by the frequency polygon.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print(LBFP)}: Print object of class \code{"LBFP"}

}}
\examples{
# Example for GLBFP function with ashua data
x <- c(200, 30)
## with a selected bin width
b <- c(0.5, 0.5)
LBFP(x, ashua[,-3], b)
## with optimal b
LBFP(x, ashua[,-3])
}
