% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBFP_estimate.R
\name{LBFP_estimate}
\alias{LBFP_estimate}
\alias{print.LBFP_estimate}
\alias{plot.LBFP_estimate}
\title{LBFP Density Estimator}
\usage{
LBFP_estimate(
  data,
  b = compute_bi_optim(data, m = rep(1, ncol(data))),
  grid_size = 20,
  grid_points = NULL,
  min_vals = apply(data, 2, min),
  max_vals = apply(data, 2, max)
)

\method{print}{LBFP_estimate}(obj)

\method{plot}{LBFP_estimate}(obj, contour = FALSE, ...)
}
\arguments{
\item{data}{Numeric matrix or data frame representing the dataset. Each row is an observation.}

\item{b}{Numeric vector representing the bin width for each dimension. Default is optimal bi calculated using \code{compute_bi_optim}.}

\item{grid_size}{Size of the grid for each dimension. Default is 20.}

\item{grid_points}{Numeric matrix of data framne as the grid to use to estimate LBFP estimator on each points of the grid. Default is NULL, in this case a grid is created from minimum to maximum on each dimension with step `grid_size`.}

\item{min_vals}{Minimum value of initial grid. Default value is the minimum on each column.}

\item{max_vals}{Maximum value of initial grid. Default value is the maximum on each column.}

\item{obj}{The list object returned by the `LBFP_estimate` function.}

\item{contour}{logical (TRUE or FALSE). \code{contour = TRUE} makes a contour plot, whereas \code{contour = FALSE} makes a 3D plot using plotly.}

\item{...}{Additional arguments for customization.}
}
\value{
A list containing the estimated density values for each point in the grid, the bin width vector `b`, the grid points and the `grid_size`.
}
\description{
This function computes the LBFP density estimate for every observation in the dataset.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print(LBFP_estimate)}: Print object of class \code{"LBFP_estimate"}

Print the density estimates calculated by the `LBFP_estimate` function.

\item \code{plot(LBFP_estimate)}: Plot object of class \code{"LBFP_estimate"}

Plot the LBFP density estimate for 1D or 2D datasets.

}}
\examples{
# Example for LBFP_estimate function with bivariate normale distribution
data <- as.data.frame(mvrnorm(n = 1000, mu =  c(0, 0), Sigma = matrix(c(1, 0.5, 0.5, 1), nrow = 2)))

##  Use of grid_size
b <- c(0.5, 0.5)
out <- LBFP_estimate(data, b, grid_size = 30)
out
plot(out)

## with a specific grid_points
grid_points <- data.frame(X = seq(-1,1,0.1), Y =  seq(-1,1,0.1))
out <- LBFP_estimate(data, b, grid_points = grid_points)
out
}
